update for soft labels: read float instead of int for labels, also works with ints as they where implicitely converted to DType in blob

+++ include/caffe/layers/image_data_layer.hpp
@@ -39,7 +39,7 @@ class ImageDataLayer : public BasePrefetchingDataLayer<Dtype> {
 
   //vector<std::pair<std::string, int> > lines_;
   //int lines_id_;
-  vector<std::pair<std::string, std::vector<int> > > lines_; 
+  vector<std::pair<std::string, std::vector<float> > > lines_;
   int lines_id_;
   int label_size_;
 };
